Check matrix against missing features.

Operate on the matrix supplied by user and check it for any features missing compared to the provided set.

Usage

check_for_missing_features(incoming_matrix, feature_set)

Arguments

incoming_matrix

The incoming matrix to be checked for any missing features.

feature_set

The feature set.

Value

matrix where column names correspond to all features from the provided set