Construct LymphGenerator matrix

Use the LymphGenerator features to construct matrix. Optionally, flatten some features having the same biological information.

Usage

classify_dlbcl_lymphgenerator(
  these_samples_metadata,
  maf_data,
  sv_data,
  seg_data,
  seq_type = "genome",
  projection = "grch37",
  output = "matrix",
  drop_after_flattening = FALSE
)

Arguments

these_samples_metadata

The metadata data frame that contains sample_id column with ids for the samples to be classified. Must also contain column pathology.

maf_data

The MAF data frame to be used for matrix assembling. At least must contain the first 45 columns of standard MAF format.

sv_data

The SV data frame to be used for matrix assembling. Must be of standard BEDPE formatting, for example, as returned by get_combined_sv.

seg_data

The SEG data frame to be used for matrix assembling. Must be of standard SEG formatting, for example, as returned by get_sample_cn_segments. Must be already adjusted for ploidy.

seq_type

String of the seq type for the sample set.

projection

String of projection of the samples. Only used to retrieve data through GAMBLR.data when it is not provided. Defaults to grch37.

output

The output to be returned. Currently only matrix is supported.

drop_after_flattening

Boolean on whether to remove features (rows) after flattening. Defaults to FALSE.

Value

binary matrix