Classify FL samples into cFL/dFL subgroups.

Use the random forest prediction model to assemble the binary matrix and use it to classify FL tummors into cFL/dFL. Please see PMID 37084389 for more details on FL genetic subgroups.

Usage

classify_fl(
  these_samples_metadata,
  maf_data,
  matrix,
  output = "predictions",
  model = RFmodel_FL
)

Arguments

these_samples_metadata

The metadata data frame that contains sample_id column with ids for the samples to be classified. Required input.

maf_data

The MAF data frame to be used for matrix assembling. The maf data must be in the grch37 projection. The chr prefix is discarded if present. Any maf columns can be provided, but the required are "Hugo_Symbol", "NCBI_Build", "Chromosome", "Start_Position", "End_Position", "Variant_Classification", "HGVSp_Short", and "Tumor_Sample_Barcode". Required input.

matrix

Optionally, if the binary feature matrix is already prepared, it can be provided in this argument.

output

The output to be returned after prediction is done. Can be one of predictions, matrix, or both. Defaults to predictions.

model

The RF model. Classifier from the paper describing cFL is used. It is not recommended to change the value of this parameter. This is a developer-only option and used for testing and functionality improvement.

Value

data frame, binary matrix, or both

Examples

if (FALSE) { # \dontrun{
meta <- get_gambl_metadata() %>%
    filter(pathology == "FL")

maf <- get_coding_ssm(
    these_samples_metadata = meta,
    tool_name = "publication"
)

classify_fl(
    these_samples_metadata = meta,
    maf_data = maf
)
classify_fl(
    these_samples_metadata = meta,
    maf_data = maf,
    output = "both"
)
} # }